Remove OfflinePDFStorageManager
