+++ Annotato/Annotato/Persistence/PersistenceManager/PersistenceManager+PDFStorageManager.swift
@@ -1,20 +1,14 @@
 import Foundation
 import AnnotatoSharedLibrary
 
-extension PersistenceManager: PDFStorageManager {
+extension PersistenceManager {
     func uploadPdf(fileSystemUrl: URL, withName name: String, completion: @escaping (Document) -> Void) {
-        OnlinePDFStorageManager(persistenceManager: self)
-            .uploadPdf(fileSystemUrl: fileSystemUrl, withName: name, completion: { _ in })
-
-        OfflinePDFStorageManager(persistenceManager: self)
+        PDFStorageManager(persistenceManager: self)
             .uploadPdf(fileSystemUrl: fileSystemUrl, withName: name, completion: completion)
     }
 
     func deletePdf(document: Document, completion: @escaping (Document) -> Void) {
-        OnlinePDFStorageManager(persistenceManager: self)
-            .deletePdf(document: document, completion: { _ in })
-
-        OfflinePDFStorageManager(persistenceManager: self)
+        PDFStorageManager(persistenceManager: self)
             .deletePdf(document: document, completion: completion)
     }
 }
